data typeof discount percentage and amount changed to decimal from int

+++ src/Common/DTO/POS/SalesCreateDto.cs
@@ -9,8 +9,8 @@ namespace POS.Common.DTO.POS
 {
     public class SalesCreateDto 
     {
-        public int DiscountPercentage { get; set; }
-        public int DiscountAmount { get; set; }
+        public decimal DiscountPercentage { get; set; }
+        public decimal DiscountAmount { get; set; }
         public int CreatedBy { get; set; }
         public List<SalesDetailCreateDto> SalesDetails { get; set; }
     }
